Split word test file.
