+++ test/extension.test.ts
@@ -3,7 +3,9 @@
 // Please refer to their documentation on https://mochajs.org/ for help.
 //
 
-import * as MotionWordTest from './Motions/Word.test';
+import * as MotionWordBasicTest from './Motions/Word/Basic.test';
+import * as MotionWordUseBlankSeparatedStyleTest from './Motions/Word/UseBlankSeparatedStyle.test';
+import * as MotionWordJapaneseTest from './Motions/Word/Japanese.test';
 import * as MotionIntegrationTest from './Motions/Integration.test';
 import * as TextObjectWordTest from './TextObjects/Word.test';
 import * as ActionSelectionTest from './Actions/Selection.test';
@@ -11,7 +13,9 @@ import * as ActionDeleteTest from './Actions/Delete.test';
 
 // Defines a Mocha test suite to group tests of similar kind together
 suite('Extension Tests', () => {
-    MotionWordTest.run();
+    MotionWordBasicTest.run();
+    MotionWordUseBlankSeparatedStyleTest.run();
+    MotionWordJapaneseTest.run();
     MotionIntegrationTest.run();
 
     ActionSelectionTest.run();
